Complete algorithms related to arrays and strings

+++ src/main/java/dprueitt/edu/max_sub_array_i/Solution.java
@@ -10,20 +10,20 @@ public class Solution
 
     public double findMaxAverage( int[] nums, int k )
     {
-        int sum = 0;
+        double sum = 0;
 
         for( int i = 0; i < k; i++ )
         {
             sum += nums[i];
         }
 
-        int ans = sum;
+        double ans = sum;
         for( int i = k; i < nums.length; i++ )
         {
             sum += nums[i] - nums[i - k];
             ans = Math.max( ans, sum );
         }
 
-        return (double) ans / (double) k;
+        return ans / k;
     }
 }
